edit/pyproject.toml
[tool.poetry]
name = "tonbtc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
django-filter = "^24.3"
drf-yasg = "^1.21.8"
pytelegrambotapi = "^4.25.0"
python-dotenv = "^1.0.1"
django-extensions = "^3.2.3"
werkzeug = "^3.1.3"
pyopenssl = "^24.3.0"
pytonapi = "<0.4.5"
pytonconnect = "^0.3.2"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-beat = "^2.7.0"
huey = "^2.5.2"
django-celery-results = "^2.5.1"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
django-debug-toolbar = "^4.4.6"
pytoniq-core = "^0.1.41"
django-advanced-filters = "^2.0.0"
django-admin-rangefilter = "^0.13.2"
tonutils = "^0.2.4"
tonsdk = "^1.0.15"
pytoniq = "^0.1.40"
aiogram = ">=2.0,<3.0"
apscheduler = "^3.11.0"
peewee = "^3.17.9"
sqlalchemy = "^2.0.38"
peewee-migrate = "^1.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[project]
name = "tonbtc"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" },
]

dependencies = [
    "django>=5.1.4,<6.0.0",
    "djangorestframework>=3.15.2,<4.0.0",
    "django-cors-headers>=4.6.0,<5.0.0",
    "django-filter>=24.3,<25.0",
    "drf-yasg>=1.21.8,<2.0.0",
    "pytelegrambotapi>=4.25.0,<5.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "django-extensions>=3.2.3,<4.0.0",
    "werkzeug>=3.1.3,<4.0.0",
    "pyopenssl>=24.3.0,<25.0.0",
    "pytonapi<0.4.5",
    "pytonconnect>=0.3.2,<0.4.0",
    "celery[redis]>=5.4.0,<6.0.0",
    "django-celery-beat>=2.7.0,<3.0.0",
    "huey>=2.5.2,<3.0.0",
    "django-celery-results>=2.5.1,<3.0.0",
    "dj-database-url>=2.3.0,<3.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "django-debug-toolbar>=4.4.6,<5.0.0",
    "pytoniq-core>=0.1.41,<0.2.0",
    "django-advanced-filters>=2.0.0,<3.0.0",
    "django-admin-rangefilter>=0.13.2,<0.14.0",
    "tonutils>=0.2.4,<0.3.0",
    "tonsdk>=1.0.15,<2.0.0",
    "pytoniq>=0.1.40,<0.2.0",
    "aiogram>=2.0,<3.0",
    "apscheduler>=3.11.0,<4.0.0",
    "peewee>=3.17.9,<4.0.0",
    "sqlalchemy>=2.0.38,<3.0.0",
    "peewee-migrate>=1.13.0,<2.0.0",
    "django-import-export>=4.3.9",
    "django-prometheus>=2.4.1",
    "djangorestframework-simplejwt>=5.5.1",
    "base58>=2.1.1",
    "solders>=0.26.0",
    "pillow>=11.3.0",
    "django-ratelimit>=4.1.0",
    "gunicorn>=23.0.0",
]

requires-python = ">=3.11"
